capstone_project/medicine_api/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
PyMySQL==1.1.1
cryptography==46.0.3

